♻️ refactor: test client file 위치 변경
