+++ apps/core/utils/__init__.py
@@ -1,5 +1,5 @@
 from apps.core.utils.create_temp_image import create_temp_image
 from apps.core.utils.s3_uploader import S3Uploader
-from apps.core.utils.test_clients import RedisTestClient
+from apps.core.tests.test_clients import RedisTestClient
 
 __all__ = ["create_temp_image", "RedisTestClient", "S3Uploader"]
