+++ apps/users/tests/test_user_email.py
@@ -9,7 +9,7 @@ from apps.core.tests.mixins.test_user_mixins import (
     IsolatedCacheTestMixin,
     VerificationMixin,
 )
-from apps.core.utils.test_clients import RedisTestClient
+from apps.core.tests.test_clients import RedisTestClient
 from apps.users.models import Withdrawals, WithdrawalsReasonChoices
 from apps.users.services.email_service import EmailVerificationService
 from apps.users.utils.enums import VerificationPurpose
